% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{A_shift_unobserved_last}
\alias{A_shift_unobserved_last}
\title{Experimental: Shift unobserved nodes in adjacency matrix (for semi-supervised estimation)}
\usage{
A_shift_unobserved_last(A, weights)
}
\arguments{
\item{A}{A square adjacency matrix representing the network}

\item{weights}{A logical vector of length equal to \code{nrow(A)} where \code{TRUE} indicates observed nodes and \code{FALSE} unobserved nodes}
}
\value{
A reordered adjacency matrix with unobserved nodes placed last
}
\description{
As part of a semi-supervised estimation strategy, this function reorders the adjacency matrix
such that observed nodes appear first and unobserved nodes (those with missing outcomes) 
are shifted to the bottom-right block. This is useful for block-decomposable estimators or 
conditional eigendecomposition techniques.
}
\note{
This function is experimental and part of a semi-supervised network estimation feature.
}
\examples{
A <- matrix(1:16, 4, 4)
weights <- c(TRUE, TRUE, FALSE, FALSE)
A_shift_unobserved_last(A, weights)

}
\seealso{
\code{\link{unobserved_shift_back}}
}
