% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{beta_sign}
\alias{beta_sign}
\title{Align signs of estimated coefficients with ground truth}
\usage{
beta_sign(beta, beta0)
}
\arguments{
\item{beta}{Estimated coefficient vector}

\item{beta0}{Ground truth or reference coefficient vector}
}
\value{
The input \code{beta} with signs aligned to \code{beta0} for the last two entries
}
\description{
Align signs of estimated coefficients with ground truth
}
\examples{
beta_sign(c(1, -2, -3), c(1, -2, 3))  # flips last entry
}
