% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{C11C21_C12C22_SuperCENT}
\alias{C11C21_C12C22_SuperCENT}
\title{Compute C11, C21, C12, C22 blocks for SuperCENT variance-covariance matrix}
\usage{
C11C21_C12C22_SuperCENT(X, u, v, Pu, Pv, betau, betav, d, n, lambda, A_perp)
}
\arguments{
\item{X}{Covariate matrix}

\item{u}{u vector}

\item{v}{v vector}

\item{Pu}{Projection matrix onto \code{u}}

\item{Pv}{Projection matrix onto \code{v}}

\item{betau}{Coefficient on centrality vector \code{u}}

\item{betav}{Coefficient on centrality vector \code{v}}

\item{d}{Leading singular value}

\item{n}{Number of nodes}

\item{lambda}{Regularization parameter}

\item{A_perp}{Residual component from multi-rank approximation (optional)}
}
\value{
A list with:
\describe{
  \item{\code{C11C21_eps}}{Influence term from response noise \eqn{\epsilon_y}}
  \item{\code{C12C22_vecE}}{Influence term from network noise \eqn{\epsilon_a}}
}
}
\description{
Compute C11, C21, C12, C22 blocks for SuperCENT variance-covariance matrix
}
\seealso{
\code{\link{C12C22_two_stage}}, \code{\link{rate_betauv_lr_2}}
}
