% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{C12C22_two_stage}
\alias{C12C22_two_stage}
\title{Compute C12 and C22 blocks for two-stage variance-covariance matrix}
\usage{
C12C22_two_stage(u, v, Pu, Pv, d, n, A_perp)
}
\arguments{
\item{u}{u vector}

\item{v}{v vector}

\item{Pu}{Projection matrix onto \code{u}, typically \code{u \%*\% t(u) / n}}

\item{Pv}{Projection matrix onto \code{v}}

\item{d}{Leading singular value}

\item{n}{Number of nodes}

\item{A_perp}{Residual adjacency matrix from multi-rank approximation (if any)}
}
\value{
A matrix representing \eqn{C_{12}} and \eqn{C_{22}} stacked vertically
}
\description{
Compute C12 and C22 blocks for two-stage variance-covariance matrix
}
\seealso{
\code{\link{C11C21_C12C22_SuperCENT}}
}
