% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{confint}
\alias{confint}
\title{Confidence interval and summary table of \eqn{\beta}}
\usage{
confint(ret, alpha = 0.05, ci = F, multi_rank = F)
}
\arguments{
\item{ret}{A  \code{two_stage}, \code{supercent}
and \code{cv.supercent} object}

\item{alpha}{The level of type-I error}

\item{ci}{If TRUE, return the confidence interval;
if FALSE, return the summary table.}
}
\value{
Output a data.frame of confidence interval
or summary table.
}
\description{
This function returns the confidence interval or 
the summary table of \code{two_stage}, \code{supercent}
and \code{cv.supercent} object.
}
