% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{confint_A}
\alias{confint_A}
\title{Confidence interval and summary table of \eqn{A}}
\usage{
confint_A(ret, alpha = 0.05, ci = F, A0 = NULL)
}
\arguments{
\item{ret}{A  \code{two_stage}, \code{supercent}
and \code{cv.supercent} object}

\item{alpha}{The level of type-I error}

\item{ci}{If TRUE, return the confidence interval;
if FALSE, return the summary table.}

\item{A0}{The true adjacency matrix}
}
\value{
Output a data.frame of confidence interval
or summary table.
}
\description{
This function returns the confidence interval or 
the summary table for the adjacency matrix of 
\code{two_stage}, \code{supercent}
and \code{cv.supercent} object.
}
