% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{epsa_hat_oracle}
\alias{epsa_hat_oracle}
\title{Legacy: Oracle estimate of network noise level}
\usage{
epsa_hat_oracle(A0, d, u, v)
}
\arguments{
\item{A0}{True adjacency matrix}

\item{d}{Leading singular value}

\item{u}{Estimated centrality vector (left)}

\item{v}{Estimated centrality vector (right)}
}
\value{
Oracle estimate of \eqn{\epsilon_a}
}
\description{
Computes \eqn{\epsilon_a} using ground truth \code{A0} and the fitted rank-1 structure from \code{u}, \code{v}, \code{d}.
}
