% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{epsy_hat}
\alias{epsy_hat}
\title{Estimate \eqn{\sigma_y}}
\usage{
epsy_hat(ret)
}
\arguments{
\item{ret}{A fitted network regression object}
}
\value{
Estimated \eqn{\epsilon_y}
}
\description{
Computes the residual standard error \eqn{\epsilon_y}.
}
\seealso{
\code{\link{epsa_hat}}, \code{\link{confint_all}}
}
