% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{func}
\alias{func}
\title{Helper: Common penalty term in SuperCENT risk expression}
\usage{
func(l, d, betau, betav, sigmay2, sigmaa2, n)
}
\arguments{
\item{l}{Regularization parameter}

\item{d}{Leading singular value}

\item{betau}{Coefficient on \code{u}}

\item{betav}{Coefficient on \code{v}}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
The combined penalty term used in risk difference expressions
}
\description{
Helper: Common penalty term in SuperCENT risk expression
}
