% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{l_optimal_lr}
\alias{l_optimal_lr}
\title{Optimal \eqn{\lambda} for SuperCENT}
\usage{
l_optimal_lr(d, beta0, sigmay2, sigmaa2, n, weights = rep(1, n))
}
\arguments{
\item{d}{Leading singular value}

\item{beta0}{True coefficient vector}

\item{sigmay2}{Regression noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Total sample size}

\item{weights}{Observation indicator vector}
}
\value{
Optimal regularization parameter \eqn{\lambda}
}
\description{
Computes the theoretical  optimal value of \eqn{\lambda}.
}
