% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_A_lr}
\alias{rate_A_lr}
\title{Expected squared error of adjacency matrix recovery under SuperCENT}
\usage{
rate_A_lr(l, u, v, beta0, d, sigmay2, sigmaa2, n)
}
\arguments{
\item{l}{Regularization parameter}

\item{u}{u vector}

\item{v}{v vector}

\item{beta0}{True coefficient vector}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
Square error
}
\description{
Expected squared error of adjacency matrix recovery under SuperCENT
}
\seealso{
\code{\link{rate_A_two_stage}}
}
