% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_A_two_stage}
\alias{rate_A_two_stage}
\title{Expected squared error of adjacency matrix recovery under two-stage}
\usage{
rate_A_two_stage(sigmaa2, n)
}
\arguments{
\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
Square error
}
\description{
Expected squared error of adjacency matrix recovery under two-stage
}
