% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_betauv_lr_2}
\alias{rate_betauv_lr_2}
\title{Rate of \code{beta_u} and \code{beta_v} of SuperCENT under rank-one}
\usage{
rate_betauv_lr_2(X, u, v, beta0, d, sigmay2, sigmaa2, n, l)
}
\arguments{
\item{X}{Covariate matrix}

\item{u}{Estimated centrality vector (left)}

\item{v}{Estimated centrality vector (right)}

\item{beta0}{True coefficient vector (with \code{beta_u}, \code{beta_v} as last entries)}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}

\item{l}{Regularization parameter}
}
\value{
Rate of \code{beta_u} and \code{beta_v} of SuperCENT
}
\description{
Rate of \code{beta_u} and \code{beta_v} of SuperCENT under rank-one
}
