% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_betauv_two_stage_check}
\alias{rate_betauv_two_stage_check}
\title{Debugging check for variance of \eqn{\\hat{\\beta}_u} in two-stage estimator}
\usage{
rate_betauv_two_stage_check(X, u, v, beta0, d, sigmay2, sigmaa2, n)
}
\arguments{
\item{X}{Covariate matrix}

\item{u}{u vector}

\item{v}{v vector}

\item{beta0}{True coefficient vector}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
A list of intermediate terms including total rate, signal and network variance terms, and key scalars
}
\description{
Debugging check for variance of \eqn{\\hat{\\beta}_u} in two-stage estimator
}
