% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_betax_lr}
\alias{rate_betax_lr}
\title{Asymptotic variance of \eqn{\\hat{\\beta}_x} under SuperCENT estimator}
\usage{
rate_betax_lr(X, u, v, beta0, d, sigmay2, sigmaa2, n)
}
\arguments{
\item{X}{Covariate matrix}

\item{u}{u vector}

\item{v}{v vector}

\item{beta0}{True coefficient vector}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
A vector of variances for each component of \eqn{\\hat{\\beta}_x}
}
\description{
Asymptotic variance of \eqn{\\hat{\\beta}_x} under SuperCENT estimator
}
\seealso{
\code{\link{rate_betax_two_stage}}
}
