% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_risk_lr}
\alias{rate_risk_lr}
\title{Asymptotic prediction risk under SuperCENT}
\usage{
rate_risk_lr(l, u, v, beta0, d, sigmay2, sigmaa2, n)
}
\arguments{
\item{l}{Regularization parameter}

\item{u}{u vector}

\item{v}{v vector}

\item{beta0}{True coefficient vector}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
Scalar risk value for SuperCENT
}
\description{
Asymptotic prediction risk under SuperCENT
}
\seealso{
\code{\link{rate_risk_two_stage}}
}
