% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_risk_two_stage}
\alias{rate_risk_two_stage}
\title{Asymptotic prediction risk under two-stage network regression}
\usage{
rate_risk_two_stage(beta0, d, sigmay2, sigmaa2, n)
}
\arguments{
\item{beta0}{True coefficient vector}

\item{d}{Leading singular value}

\item{sigmay2}{Outcome noise variance}

\item{sigmaa2}{Network noise variance}

\item{n}{Sample size}
}
\value{
Risk value
}
\description{
Asymptotic prediction risk under two-stage network regression
}
