% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_u_lr}
\alias{rate_u_lr}
\title{Rate of centrality \eqn{\hat{u}} of SuperCENT under rank-one model}
\usage{
rate_u_lr(l, d, beta0, sigmay2, sigmaa2, n)
}
\arguments{
\item{l}{Regularization parameter}

\item{d}{Leading singular value}

\item{beta0}{True coefficient vector, with \code{beta_u} and \code{beta_v} as last two entries}

\item{sigmay2}{Variance of outcome noise}

\item{sigmaa2}{Variance of network noise}

\item{n}{Sample size}
}
\value{
SuperCENT rate of \eqn{\hat{u}}
}
\description{
Rate of centrality \eqn{\hat{u}} of SuperCENT under rank-one model
}
