% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rate_u_two_stage}
\alias{rate_u_two_stage}
\title{Rate of \eqn{\hat{u}} of two-stage estimator under rank-one model}
\usage{
rate_u_two_stage(d, sigmaa2, n)
}
\arguments{
\item{d}{Leading singular value}

\item{sigmaa2}{Variance of network noise}

\item{n}{Sample size}
}
\value{
Two-stage rate of \eqn{\hat{u}}
}
\description{
Rate of \eqn{\hat{u}} of two-stage estimator under rank-one model
}
