% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ret_constructor}
\alias{ret_constructor}
\title{Initialize a result object for network regression models}
\usage{
ret_constructor(A, X, y)
}
\arguments{
\item{A}{The adjacency matrix}

\item{X}{The design matrix}

\item{y}{The response vector}
}
\value{
A list with initialized slots for centrality estimation, regression coefficients, and meta info
}
\description{
Constructs an empty list with initialized slots for use in centrality-based regression procedures.
}
\examples{
ret <- ret_constructor(matrix(0, 2, 2), matrix(1, 2, 1), c(1, 2))
}
