% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ret_sign}
\alias{ret_sign}
\title{Align signs of full network regression output with ground truth}
\usage{
ret_sign(ret, beta0vec, u0, v0)
}
\arguments{
\item{ret}{A result object from \code{supercent}, \code{two_stage}, or similar}

\item{beta0vec}{Reference coefficient vector}

\item{u0}{Reference left latent vector}

\item{v0}{Reference right latent vector}
}
\value{
The input \code{ret} with aligned \code{u}, \code{v}, \code{beta}, and optionally \code{beta_hat}
}
\description{
Align signs of full network regression output with ground truth
}
\seealso{
\code{\link{beta_sign}}, \code{\link{u_sign}}, \code{\link{adjust_sign}}
}
