\name{SuperCENT-package}
\alias{SuperCENT-package}
\alias{SuperCENT}
\docType{package}
\title{
\packageTitle{SuperCENT}
}
\description{
\packageDescription{SuperCENT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SuperCENT}
\packageIndices{SuperCENT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SuperCENT}

Maintainer: \packageMaintainer{SuperCENT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
