% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_sign}
\alias{u_sign}
\title{Align sign of estimated centrality with ground truth}
\usage{
u_sign(u, u0)
}
\arguments{
\item{u}{Estimated latent vector}

\item{u0}{Ground truth or reference latent vector}
}
\value{
The input \code{u} with aligned sign
}
\description{
Align sign of estimated centrality with ground truth
}
\examples{
u_sign(c(-1, -2), c(1, 2))  # flips sign
}
