% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_beta}
\alias{var_beta}
\title{Variance of \eqn{hat{\beta_u}} and \eqn{hat{\beta_v}} ordered by column major}
\usage{
var_beta(
  X,
  u,
  v,
  beta0,
  d,
  sigmay2,
  sigmaa2,
  n,
  output = "uvX",
  method = "two_stage",
  verbose = F,
  lambda = NA,
  multi_rank = F,
  multi_rank_ = NULL,
  A_perp = NULL
)
}
\arguments{
\item{X}{design matrix X}

\item{u}{u vector}

\item{v}{v vector}

\item{beta0}{beta vector}

\item{d}{d}

\item{sigmay2}{sigma_y^2}

\item{sigmaa2}{sigma_a^2}

\item{n}{number of nodes}

\item{output}{"uv", "u", "v" to select output for \eqn{beta_u} and/or \eqn{beta_v}}

\item{verbose}{print first and second term of the rate}
}
\value{
A vector of rariance for \eqn{hat{\beta_u}} and \eqn{hat{\beta_v}}
}
\description{
Variance of \eqn{hat{\beta_u}} and \eqn{hat{\beta_v}} ordered by column major
}
