% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_mat}
\alias{var_mat}
\title{Variance estimation for network regression parameters}
\usage{
var_mat(
  X_train,
  beta_u,
  beta_v,
  d,
  u,
  v,
  epsa2,
  epsy2,
  l = NULL,
  method = "lr"
)
}
\arguments{
\item{X_train}{Covariate matrix for observed nodes}

\item{beta_u}{Estimated coefficient on centrality \code{u}}

\item{beta_v}{Estimated coefficient on centrality \code{v}}

\item{d}{Estimated leading singular value}

\item{u}{Estimated centrality vector (left)}

\item{v}{Estimated centrality vector (right)}

\item{epsa2}{Estimated noise variance in the network (\code{A})}

\item{epsy2}{Estimated noise variance in the outcome (\code{y})}

\item{l}{Regularization parameter (optional, used in low-rank models)}

\item{method}{Estimation method, one of \code{"lr"} ) (SuperCENT) or \code{"two_stage"}}
}
\value{
A numeric vector of estimated variances for \code{u}, \code{v}, \code{beta_u}, \code{beta_v}, and \code{beta_x}
}
\description{
Variance estimation for network regression parameters
}
