% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_mat_A_supercent}
\alias{var_mat_A_supercent}
\title{Variance of \eqn{hat{a}_{ij}} ordered by column major}
\usage{
var_mat_A_supercent(X, l, d, beta0, sigmay2, sigmaa2, u, v, n)
}
\arguments{
\item{X}{design matrix X}

\item{l}{tuning parameter}

\item{sigmay2}{sigma_y^2}

\item{sigmaa2}{sigma_a^2}

\item{u}{u vector}

\item{v}{v vector}

\item{n}{number of nodes}
}
\value{
A vector of variance of \eqn{hat{a}_{ij}}
}
\description{
Variance of \eqn{hat{a}_{ij}} ordered by column major
}
