% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_mat_A_two_stage}
\alias{var_mat_A_two_stage}
\title{Variance of \eqn{hat{a}^{ts}_{ij}} ordered by column major}
\usage{
var_mat_A_two_stage(sigmaa2, u, v, n)
}
\arguments{
\item{sigmaa2}{sigma_a^2}

\item{u}{u vector}

\item{v}{v vector}

\item{n}{number of nodes}
}
\value{
A vector of variance of \eqn{hat{a}^{ts}_{ij}}
}
\description{
Variance of \eqn{hat{a}^{ts}_{ij}} ordered by column major
}
