% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vec_norm}
\alias{vec_norm}
\title{Compute the Euclidean norm of a vector}
\usage{
vec_norm(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The Euclidean norm \eqn{\|x\| = \sqrt{\sum x_i^2}}
}
\description{
Compute the Euclidean norm of a vector
}
\examples{
vec_norm(c(3, 4))  # returns 5
}
