% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_mat}
\alias{var_mat}
\title{Calculate variance}
\usage{
var_mat(
  X_train,
  beta_u,
  beta_v,
  d,
  u,
  v,
  epsa2,
  epsy2,
  l = NULL,
  method = "lr"
)
}
\arguments{
\item{X_train}{X train matrix}

\item{method}{method}
}
\value{
variance of hat u,v,betax, betau, betav
}
\description{
Calculate variance
}
